+++ CHANGELOG.md
@@ -38,6 +38,7 @@
 - [#8897](https://github.com/influxdata/influxdb/pull/8897): Add message pack format for query responses.
 - [#8886](https://github.com/influxdata/influxdb/pull/8886): Improved compaction scheduling
 - [#8690](https://github.com/influxdata/influxdb/issues/8690): Implicitly decide on a lower limit for fill queries when none is present.
+- [#7355](https://github.com/influxdata/influxdb/issues/7355): Create a command to truncated shard groups 
 
 ### Bugfixes